requirements.txt
crewai>=0.28.0
crewai-tools>=0.4.0
pydantic>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
streamlit>=1.28.0
requests>=2.31.0
langchain>=0.1.0
langchain-openai>=0.0.5
